requirements.txt
PyRSMQ==0.4.2